+++ CHANGELOG.rst
@@ -10,6 +10,8 @@ Unreleased
 Added
 ~~~~~
 
+* inject-server: adding httpPostAdmin test method. ``PHAB_ID=D482624``
+
 Changed
 ~~~~~~~
 
@@ -35,6 +37,7 @@ Added
 * finatra-examples: Ensure there are Java and Scala examples for the different
   types of applications and servers which can be built with Finatra. Update `/examples`
   directory layout for discoverability and consistency. ``PHAB_ID=D469677``
+
 
 Changed
 ~~~~~~~